+++ Packages/BrandonUtils/Standalone/Collections/IPrimaryKeyed.cs
@@ -1,4 +1,7 @@
-﻿namespace BrandonUtils.Standalone.Collections {
+﻿using JetBrains.Annotations;
+
+namespace BrandonUtils.Standalone.Collections {
+    [PublicAPI]
     public interface IPrimaryKeyed<out T> {
         T PrimaryKey { get; }
     }