+++ Packages/BrandonUtils/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.15.1] - 9/2/2021
+
+### Changed
+
+- Renamed `KeyedList` -> `PrimaryKeyedList`
+- `KeyedList` is now a `KeyedCollection` implementation that doesn't require `IPrimaryKeyed` members
+
 ## [2.14.3] - 9/2/2021
 
 ### Fixed